requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.1
django-crispy-forms==1.9.2
djangorestframework==3.11.1
joblib @ file:///tmp/build/80754af9/joblib_1594236160679/work
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sklearn-pandas==2.0.1
sqlparse==0.3.1
threadpoolctl==2.1.0
wincertstore==0.2
